feat (matrix): nueva función para comparar matrices, retorna 1 si son iguales; 0 en caso contrario. - Debido a que los datos dentro de las matrices son de tipo double, se necesita una tolerancia absoluta y relativa - Analogía simple Error de 1 dólar: en 10 dólares → mucho (10%) en 1,000,000 dólares → insignificante

+++ test/test_matrix.c
@@ -68,7 +68,7 @@ void test_random_matrix_range() {
 void test_zero_matrix() {
     int n = 4;
 
-    int** M = allocMatrix(n);
+    double** M = allocMatrix(n);
 
     // llenar con basura
     for (int i = 0; i < n; i++)
